[add] add lib.rs and move `EmulateExtension` trait to it
