+++ src/trap/exception/instruction_handler.rs
@@ -5,10 +5,10 @@
 
 use super::hs_forward_exception;
 use crate::emulate_extension::zicfiss::ZICFISS_DATA;
-use crate::emulate_extension::EmulateExtension;
 use crate::HYPERVISOR_DATA;
 
 use core::arch::asm;
+use hikami::EmulateExtension;
 use raki::{Instruction, OpcodeKind};
 use riscv::register::{sepc, stval};
